+++ src/Walk/WriteRules.php
@@ -6,7 +6,8 @@ namespace JonBaldie\ExplicitnessChecker\Walk;
 
 /**
  * The rules for nodes that write to some of their children: assignments,
- * increments and decrements, foreach targets, catch variables and unset.
+ * increments and decrements, foreach targets, catch variables, unset and
+ * by-reference arguments to built-ins.
  */
 class WriteRules extends RuleChain
 {
@@ -20,6 +21,7 @@ class WriteRules extends RuleChain
             new ForeachRule(),
             new CatchRule(),
             new UnsetRule(),
+            new ByReferenceCallRule(),
         ];
     }
 }